+++ ssstar/src/credentials_provider.rs
@@ -10,7 +10,7 @@ use std::{fmt::Debug, sync::Arc};
 
 /// This represents the credentials to access to AWS API.
 #[derive(Clone, Debug)]
-pub struct CustomCredentialsUpdateOutput {
+pub struct CredentialsUpdateOutput {
     pub access_key: String,
     pub secret_key: String,
     pub session_token: Option<String>,
@@ -18,34 +18,34 @@ pub struct CustomCredentialsUpdateOutput {
 
 /// This is a trait that represents a callback to update credentials when they become expired
 #[async_trait]
-pub trait CustomCredentialsUpdateCallback: Debug + Send + Sync {
+pub trait CredentialsUpdateCallback: Debug + Send + Sync {
     /// This is not called for each AWS API call, this is called only when the
     /// current credentials are expired
     async fn update_credentials(
         &self,
-    ) -> Result<CustomCredentialsUpdateOutput, Box<dyn std::error::Error + Send + Sync + 'static>>;
+    ) -> Result<CredentialsUpdateOutput, Box<dyn std::error::Error + Send + Sync + 'static>>;
 }
 
 /// Custom credentials provider that actually implements `ProvideCredentials` and implements the needed
 /// traits to make it possible to put it into config as an option
 #[derive(Debug, Clone)]
-pub struct CustomCredentialsProvider(Arc<dyn CustomCredentialsUpdateCallback>);
+pub struct CredentialsProvider(Arc<dyn CredentialsUpdateCallback>);
 
 /// This is needed to make it possible put it in clap configuration, the implementation doesn't matter
 /// because this must be never called
-impl PartialEq for CustomCredentialsProvider {
+impl PartialEq for CredentialsProvider {
     fn eq(&self, _other: &Self) -> bool {
         unreachable!("BUG: You should never compare `CustomCredentialsProvider`");
     }
 }
 
-impl CustomCredentialsProvider {
-    pub fn new(inner: impl CustomCredentialsUpdateCallback + 'static) -> Self {
+impl CredentialsProvider {
+    pub fn new(inner: impl CredentialsUpdateCallback + 'static) -> Self {
         Self(Arc::new(inner))
     }
 }
 
-impl ProvideCredentials for CustomCredentialsProvider {
+impl ProvideCredentials for CredentialsProvider {
     fn provide_credentials<'a>(
         &'a self,
     ) -> aws_credential_types::provider::future::ProvideCredentials<'a>